🔧 fix: Adjustments made according to Code Review requirements
